Renamed apolloManager to apollo.
